% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBSEERBound.R
\docType{data}
\name{BBSEERBound}
\alias{BBSEERBound}
\title{Biscayne Bay and Southeastern Everglades Ecosystem Restoration Project Boundary}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 5 columns.
}
\source{
USACE
}
\usage{
data(BBSEERBound)
}
\description{
For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: USACE - WERP https://www.saj.usace.army.mil/BBSEER/
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
library(sf)
data(BBSEERBound)
plot(st_geometry(BBSEERBound))
}
\keyword{dataset}
