% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canals.R
\docType{data}
\name{canals}
\alias{canals}
\title{South Florida Water Management District Canals}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 2723 rows and 18 columns.
}
\source{
FCE LTER
}
\usage{
data(canals)
}
\description{
Simplified South Florida Water Management District Canals

For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: FCE LTER repo https://github.com/FCE-LTER/maps_FCE_sites_grayscale_base/tree/main
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(canals)
plot(st_geometry(canals))
}
}
\keyword{dataset}
