% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEPP.R
\docType{data}
\name{CEPP}
\alias{CEPP}
\title{Central Everglades Planning}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 3 columns.
}
\source{
PJ
}
\usage{
data(CEPP)
}
\description{
For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: Conceptually USACE, digitized by P Julian
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
library(sf)
data(CEPP)
plot(st_geometry(CEPP))
}
\keyword{dataset}
