% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties_fl.R
\docType{data}
\name{counties_fl}
\alias{counties_fl}
\title{Florida County Boundaries}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 67 rows and 16 columns.
}
\source{
USAboundaries
}
\usage{
data(counties_fl)
}
\description{
Florida county boundary shapefile from \code{USAboundaries} R-package

For purposes of this package, data is reprojected from WGS84 to UTM17N.

\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(counties_fl)
plot(st_geometry(counties_fl))
}
}
\keyword{dataset}
