% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaa.R
\docType{data}
\name{eaa}
\alias{eaa}
\title{SFWMD Everglades Agricultural Area (EAA) Boundary}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 3 rows and 8 columns.
}
\source{
SFWMD
}
\usage{
data(eaa)
}
\description{
The generlized spatial extents (boundary) of the "Everglades Agricultural Area" or "EAA" as describe in Florida Statute 373.4592

For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: SFWMD https://geo-sfwmd.hub.arcgis.com/datasets/sfwmd::everglades-agricultural-area-eaa-boundary/about
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(eaa)
plot(st_geometry(eaa))
}
}
\keyword{dataset}
