% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvPA.R
\docType{data}
\name{EvPA}
\alias{EvPA}
\title{Everglades Protection Area Boundary}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 5 columns.
}
\source{
SFWMD
}
\usage{
data(EvPA)
}
\description{
As defined in 40E-63 FAC. Shapefiles have been merged into a single boundary file.

For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: SFWMD https://geo-sfwmd.hub.arcgis.com/datasets/78f52a0996a14d55ac411dc7ee2ec002_0/explore?location=25.755687\%2C-80.870991\%2C8.68
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
library(sf)
data(EvPA)
plot(st_geometry(EvPA))
}
\keyword{dataset}
