% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCELTER_sites.R
\docType{data}
\name{FCELTER_sites}
\alias{FCELTER_sites}
\title{FCE LTER long-term monitoring sites}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 18 rows and 12 columns.
}
\source{
FCE LTER
}
\usage{
data(FCELTER_sites)
}
\description{
Long term monitoring locaiton within Shark River Slough, Taylor Slough and Florida Bay

Data Source: FCE LTER https://fce-lter.fiu.edu/data/GIS/
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(FCELTER_sites)
plot(st_geometry(FCELTER_sites))
}
}
\keyword{dataset}
