% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAB.R
\docType{data}
\name{FLAB}
\alias{FLAB}
\title{Florida Bay FATHOM Basins}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 56 rows and 5 columns.
}
\source{
DOI
}
\usage{
data(FLAB)
}
\description{
Basins used in the FATHOM model
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
library(sf)
data(FLAB)
plot(st_geometry(FLAB))
}
\keyword{dataset}
