% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWCShore_clip.R
\docType{data}
\name{FWCShore_clip}
\alias{FWCShore_clip}
\title{Florida Shoreline 1:12000}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 7645 rows and 7 columns.
}
\source{
FWC
}
\usage{
data(FWCShore_clip)
}
\description{
Florida'a shoreline developed by the Florida Fish and Wildlife Conservation Commission and clipped to the SFWMD (with 20km buffer).

For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: FL FWC https://geodata.myfwc.com/datasets/myfwc::florida-shoreline-1-to-12000-scale/about
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(FWCShore_clip)
plot(st_geometry(FWCShore_clip))
}
}
\keyword{dataset}
