% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakes.R
\docType{data}
\name{lakes}
\alias{lakes}
\title{Lakes}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 75 rows and 17 columns.
}
\source{
SFWMD
}
\usage{
data(lakes)
}
\description{
Lakes from the South Florida Water Management District AHED Waterbodies shapefile.

For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: SFWMD https://geodata.floridagio.gov/datasets/46a54d27c6bb4779bba9e87399eb343f_0/about
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(lakes)
plot(st_geometry(lakes))
}
}
\keyword{dataset}
