% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOK.R
\docType{data}
\name{LOK}
\alias{LOK}
\title{Generalized Lake Okeechobee Boundary}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 17 columns.
}
\source{
SFWMD
}
\usage{
data(LOK)
}
\description{
For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: SFWMD
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(LOK)
plot(st_geometry(LOK))
}
}
\keyword{dataset}
