% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOKLitt.R
\docType{data}
\name{LOKLitt}
\alias{LOKLitt}
\title{Generalized Lake Okeechobee Littoral Zone Boundary}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\source{
SFWMD
}
\usage{
data(LOKLitt)
}
\description{
Generalized extent of the littoral zone in Lake Okeechobee based on vegetation surveys
For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: SFWMD
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(LOKLitt)
plot(st_geometry(LOKLitt))
}
}
\keyword{dataset}
