% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOSOM.R
\docType{data}
\name{LOSOM}
\alias{LOSOM}
\title{Lake Okeechobee System Operating Manual}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\source{
USACE & PJ
}
\usage{
data(LOSOM)
}
\description{
For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: USACE and redigitizing by P Julian
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
library(sf)
data(LOSOM)
plot(st_geometry(LOSOM))
}
\keyword{dataset}
