% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lostman_slough.R
\docType{data}
\name{lostman_slough}
\alias{lostman_slough}
\title{Generalized outline of Lostmans Slough}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
data(lostman_slough)
}
\description{
Generalized outline of Lostmans Slough based on vegetation, topography and McVoy et al. 2011
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
library(sf)
data(lostman_slough)
plot(st_geometry(lostman_slough))
}
\keyword{dataset}
