% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOWRP.R
\docType{data}
\name{LOWRP}
\alias{LOWRP}
\title{Lake Okeechobee Watershed Restoration Project}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 3 columns.
}
\source{
USACE + PJ
}
\usage{
data(LOWRP)
}
\description{
For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: USACE and redigitized by P Julian
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
library(sf)
data(LOWRP)
plot(st_geometry(LOWRP))
}
\keyword{dataset}
