% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_EVERSpat.R
\name{make_EVERSpat}
\alias{make_EVERSpat}
\title{List of data in the EVERSpatDat library}
\usage{
make_EVERSpat()
}
\value{
list of available Everglades related shapefiles.
}
\description{
List of data in the EVERSpatDat library
}
\examples{
make_EVERSpat()
}
