% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps_clipped.R
\docType{data}
\name{nps_clipped}
\alias{nps_clipped}
\title{National Park Service areas in Everglades region}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 21 columns.
}
\source{
NPS
}
\usage{
data(nps_clipped)
}
\description{
For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: NPS https://public-nps.opendata.arcgis.com/datasets/nps-boundary-1/explore
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(nps_clipped)
plot(st_geometry(nps_clipped))
}
}
\keyword{dataset}
