% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_EVERSpat.R
\name{print.EVERSpat}
\alias{print.EVERSpat}
\title{print a EVERSpat object}
\usage{
\method{print}{EVERSpat}(x, ...)
}
\arguments{
\item{x}{EVERSpat object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print a EVERSpat object
}
