% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaltExtent.R
\docType{data}
\name{SaltExtent}
\alias{SaltExtent}
\title{Inland Extent of Saltwater (2018)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 9 rows and 10 columns.
}
\source{
USGS Caribbean-Florida Water Science Center
}
\usage{
data(SaltExtent)
}
\description{
The isoline indicating the extent of saltwater intrusion (as detected in 2018) within Miami-Dade County

Data Source: USGS Caribbean-Florida Water Science Center, https://www.sciencebase.gov/catalog/item/5cd0a9ede4b09b8c0b79a53b
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(SaltExtent)
plot(st_geometry(SaltExtent))
}
}
\keyword{dataset}
