% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfwmd_bound.R
\docType{data}
\name{sfwmd_bound}
\alias{sfwmd_bound}
\title{South Florida Water Management District Boundary}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\source{
SFWMD
}
\usage{
data(sfwmd_bound)
}
\description{
South Florida Water Management District Boundary. This shapefile was queried from a large polygon shapefile that contained the boundary of all Water Management Districts wihtin Florida

For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: SFWMD https://geo-sfwmd.hub.arcgis.com/datasets/sfwmd::water-management-district-boundaries/about
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(sfwmd_bound)
plot(st_geometry(sfwmd_bound))
}
}
\keyword{dataset}
