% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloughs.R
\docType{data}
\name{sloughs}
\alias{sloughs}
\title{General delineation of Shark River and Taylor Sloughs within Everglades National Park}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 2 rows and 4 columns.
}
\source{
FCE LTER
}
\usage{
data(sloughs)
}
\description{
General delineation of the two major drainages within Everglades National Park

Data Source: FCE LTER https://fce-lter.fiu.edu/data/GIS/
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(sloughs)
plot(st_geometry(sloughs))
}
}
\keyword{dataset}
