% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TribalAreas.R
\docType{data}
\name{TribalAreas}
\alias{TribalAreas}
\title{Florida Tribal Areas}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 9 rows and 24 columns.
}
\source{
TIGER
}
\usage{
data(TribalAreas)
}
\description{
A subset of the Tribal lands as identified in the TIGER American Indian Lands 2020.

For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: TIGER https://geodata.floridagio.gov/datasets/9b310e43e31d498098b6ef0e47534800/explore
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(TribalAreas)
plot(st_geometry(TribalAreas))
}
}
\keyword{dataset}
