% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTBrigde.R
\docType{data}
\name{TTBridge}
\alias{TTBridge}
\title{Tamiami Trail Bridges}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 3 rows and 11 columns.
}
\source{
FCE LTER
}
\usage{
data(TTBridge)
}
\description{
Three line segements identifing the 2.5 mile and 1 mile bridges along US-41/Tamiami Trail north of Everlgades National Park

Data Source: FCE LTER repo https://github.com/FCE-LTER/maps_FCE_sites_grayscale_base
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(TTBridge)
plot(st_geometry(TTBridge))
}
}
\keyword{dataset}
