% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US41_US1.R
\docType{data}
\name{US41_US1}
\alias{US41_US1}
\title{Major Roads for FCE LTER}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 133 rows and 23 columns.
}
\source{
FCE US41_US1
}
\usage{
data(US41_US1)
}
\description{
Data Source: FCE LTER https://fce-lter.fiu.edu/data/GIS/
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(US41_US1)
plot(st_geometry(US41_US1))
}
}
\keyword{dataset}
