% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCAs.R
\docType{data}
\name{WCAs}
\alias{WCAs}
\title{SFWMD Water Conservation Areas}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 5 rows and 14 columns.
}
\source{
SFWMD
}
\usage{
data(WCAs)
}
\description{
A subset of the SFWMD watersheds file to only include the Water Conservation Areas

For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: SFWMD https://geo-sfwmd.hub.arcgis.com/datasets/sfwmd::ahed-watersheds/
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
\dontrun{
library(sf)
data(WCAs)
plot(st_geometry(WCAs))
}
}
\keyword{dataset}
