% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WERPBound.R
\docType{data}
\name{WERPBound}
\alias{WERPBound}
\title{Western Everglades Restoration Project Boundary}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 12 columns.
}
\source{
USACE
}
\usage{
data(WERPBound)
}
\description{
For purposes of this package, data is reprojected from WGS84 to UTM17N.

Data Source: USACE - WERP https://www.saj.usace.army.mil/WERP/
\describe{
\item{epsg}{EPSG:26917; NAD83 UTM17N}
}
}
\examples{
library(sf)
data(WERPBound)
plot(st_geometry(WERPBound))
}
\keyword{dataset}
